Improve aura node beam texture and cleanup rendering code
